Move TTargetTransfer to target_transfer.h (#15269)
