+++ ydb/core/tx/replication/controller/event_util.cpp
@@ -1,5 +1,5 @@
 #include "event_util.h"
-#include "target_table.h"
+#include "target_transfer.h"
 
 namespace NKikimr::NReplication::NController {
