+++ ydb/core/tx/replication/controller/replication.cpp
@@ -4,6 +4,7 @@
 #include "secret_resolver.h"
 #include "target_discoverer.h"
 #include "target_table.h"
+#include "target_transfer.h"
 #include "tenant_resolver.h"
 #include "util.h"
