+++ ydb/core/tx/replication/controller/target_discoverer.cpp
@@ -2,6 +2,7 @@
 #include "private_events.h"
 #include "target_discoverer.h"
 #include "target_table.h"
+#include "target_transfer.h"
 #include "util.h"
 
 #include <ydb/core/base/path.h>
