+++ ydb/core/tx/replication/controller/tx_discovery_targets_result.cpp
@@ -1,6 +1,5 @@
 #include "controller_impl.h"
-#include "target_table.h"
-#include "util.h"
+#include "target_transfer.h"
 
 #include <util/string/join.h>
