+++ ydb/core/tx/replication/controller/tx_init.cpp
@@ -1,5 +1,6 @@
 #include "controller_impl.h"
 #include "target_table.h"
+#include "target_transfer.h"
 
 namespace NKikimr::NReplication::NController {
